# docs/generated/api/swagger.yaml
definitions:
...
  handlers.ErrorResponse:
    properties:
      code:
        type: integer
      error:
        type: string
      message:
        type: string
    type: object
  handlers.PaginatedResponse:
    properties:
      data: {}
      page:
        type: integer
      page_size:
        type: integer
      pages:
        type: integer
      total:
        type: integer
    type: object
...
  models.CustomField:
    properties:
      active:
        type: boolean
      created_at:
        type: string
      id:
        type: integer
      key:
        description: stable identifier (slug)
        type: string
      name:
        type: string
      options_json:
        description: JSON array (for select/multiselect)
        type: string
      required:
        type: boolean
      resource:
        description: ticket
        type: string
      show_when_json:
        description: JSON condition expression
        type: string
      tenant_id:
        type: string
      type:
        description: string, number, boolean, date, select, multiselect
        type: string
      updated_at:
        type: string
      validation_json:
        description: JSON object (min/max/regex/etc)
        type: string
      workspace_id:
        type: string
    type: object
...
  models.Message:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      sender:
        description: user, ai, agent
        type: string
      session:
        $ref: '#/definitions/models.Session'
      session_id:
        type: string
      tenant_id:
        type: string
      type:
        description: text, image, file, system
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: integer
      workspace_id:
        type: string
    type: object
...
  models.Session:
    properties:
      agent:
        $ref: '#/definitions/models.User'
      agent_id:
        type: integer
      created_at:
        type: string
      ended_at:
        type: string
      id:
        type: string
      messages:
        items:
          $ref: '#/definitions/models.Message'
        type: array
      platform:
        description: web, telegram, wechat, etc.
        type: string
      started_at:
        type: string
      status:
        description: active, ended, transferred
        type: string
      tenant_id:
        type: string
      ticket:
        $ref: '#/definitions/models.Ticket'
      ticket_id:
        type: integer
      updated_at:
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: integer
      workspace_id:
        type: string
    type: object
  models.Ticket:
    properties:
      agent:
        $ref: '#/definitions/models.User'
      agent_id:
        type: integer
      attachments:
        items:
          $ref: '#/definitions/models.TicketFile'
        type: array
      category:
        description: technical, billing, general, complaint
        type: string
      closed_at:
        type: string
      comments:
        items:
          $ref: '#/definitions/models.TicketComment'
        type: array
      created_at:
        type: string
      custom_field_values:
        items:
          $ref: '#/definitions/models.TicketCustomFieldValue'
        type: array
      customer:
        allOf:
        - $ref: '#/definitions/models.User'
        description: 关联关系
      customer_id:
        type: integer
      description:
        type: string
      due_date:
        type: string
      id:
        type: integer
      priority:
        description: low, normal, high, urgent
        type: string
      resolved_at:
        type: string
      session:
        $ref: '#/definitions/models.Session'
      session_id:
        type: string
      source:
        description: web, email, phone, chat
        type: string
      status:
        description: open, assigned, in_progress, resolved, closed
        type: string
      status_history:
        items:
          $ref: '#/definitions/models.TicketStatus'
        type: array
      tags:
        description: 标签，逗号分隔
        type: string
      tenant_id:
        type: string
      title:
        type: string
      updated_at:
        type: string
      workspace_id:
        type: string
    type: object
  models.TicketComment:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      ticket:
        $ref: '#/definitions/models.Ticket'
      ticket_id:
        type: integer
      type:
        description: comment, internal_note, system
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: integer
    type: object
  models.TicketCustomFieldValue:
    properties:
      created_at:
        type: string
      custom_field:
        $ref: '#/definitions/models.CustomField'
      custom_field_id:
        type: integer
      id:
        type: integer
      ticket_id:
        type: integer
      updated_at:
        type: string
      value:
        type: string
    type: object
  models.TicketFile:
    properties:
      created_at:
        type: string
      file_name:
        type: string
      file_path:
        type: string
      file_size:
        type: integer
      id:
        type: integer
      mime_type:
        type: string
      ticket:
        $ref: '#/definitions/models.Ticket'
      ticket_id:
        type: integer
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: integer
    type: object
  models.TicketStatus:
    properties:
      created_at:
        type: string
      from_status:
        type: string
      id:
        type: integer
      reason:
        type: string
      ticket:
        $ref: '#/definitions/models.Ticket'
      ticket_id:
        type: integer
      to_status:
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: integer
    type: object
...
  models.User:
    properties:
      avatar:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      last_login:
        type: string
      name:
        type: string
      phone:
        type: string
      role:
        description: customer, agent, admin
        type: string
      sessions:
        description: 关联关系
        items:
          $ref: '#/definitions/models.Session'
        type: array
      status:
        description: active, inactive, banned
        type: string
      tickets:
        items:
          $ref: '#/definitions/models.Ticket'
        type: array
      token_valid_after:
        type: string
      token_version:
        type: integer
      updated_at:
        type: string
      username:
        type: string
    type: object
...
  services.CustomerActivity:
    properties:
      customer_id:
        type: integer
      recent_messages:
        items:
          $ref: '#/definitions/models.Message'
        type: array
      recent_sessions:
        items:
          $ref: '#/definitions/models.Session'
        type: array
      recent_tickets:
        items:
          $ref: '#/definitions/models.Ticket'
        type: array
    type: object
  services.CustomerCreateRequest:
    properties:
      company:
        type: string
      email:
        type: string
      industry:
        type: string
      name:
        type: string
      notes:
        type: string
      phone:
        type: string
      priority:
        type: string
      source:
        type: string
      tags:
        type: string
      username:
        type: string
    required:
    - email
    - username
    type: object
  services.CustomerIndustryCount:
    properties:
      count:
        type: integer
      industry:
        type: string
    type: object
  services.CustomerInfo:
    properties:
      avatar:
        type: string
      company:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      industry:
        type: string
      last_login:
        type: string
      name:
        type: string
      notes:
        type: string
      phone:
        type: string
      priority:
        type: string
      role:
        description: customer, agent, admin
        type: string
      sessions:
        description: 关联关系
        items:
          $ref: '#/definitions/models.Session'
        type: array
      source:
        type: string
      status:
        description: active, inactive, banned
        type: string
      tags:
        type: string
      tickets:
        items:
          $ref: '#/definitions/models.Ticket'
        type: array
      token_valid_after:
        type: string
      token_version:
        type: integer
      updated_at:
        type: string
      username:
        type: string
    type: object
  services.CustomerPriorityCount:
    properties:
      count:
        type: integer
      priority:
        type: string
    type: object
  services.CustomerSourceCount:
    properties:
      count:
        type: integer
      source:
        type: string
    type: object
  services.CustomerStats:
    properties:
      active:
        type: integer
      by_industry:
        items:
          $ref: '#/definitions/services.CustomerIndustryCount'
        type: array
      by_priority:
        items:
          $ref: '#/definitions/services.CustomerPriorityCount'
        type: array
      by_source:
        items:
          $ref: '#/definitions/services.CustomerSourceCount'
        type: array
      new_this_week:
        type: integer
      total:
        type: integer
    type: object
  services.CustomerUpdateRequest:
    properties:
      company:
        type: string
      industry:
        type: string
      name:
        type: string
      notes:
        type: string
      phone:
        type: string
      priority:
        type: string
      source:
        type: string
      status:
        type: string
      tags:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/customers:
    get:
      consumes:
      - application/json
      description: 获取客户列表，支持分页和过滤
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: page_size
        type: integer
      - description: 搜索关键词
        in: query
        name: search
        type: string
      - collectionFormat: csv
        description: 行业过滤
        in: query
        items:
          type: string
        name: industry
        type: array
      - collectionFormat: csv
        description: 来源过滤
        in: query
        items:
          type: string
        name: source
        type: array
      - collectionFormat: csv
        description: 优先级过滤
        in: query
        items:
          type: string
        name: priority
        type: array
      - collectionFormat: csv
        description: 状态过滤
        in: query
        items:
          type: string
        name: status
        type: array
      - description: 标签过滤
        in: query
        name: tags
        type: string
      - description: 排序字段
        in: query
        name: sort_by
        type: string
      - description: 排序方向
        in: query
        name: sort_order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/services.CustomerInfo'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: 获取客户列表
      tags:
      - 客户管理
    post:
      consumes:
      - application/json
      description: 创建新的客户账户
      parameters:
      - description: 客户信息
        in: body
        name: customer
        required: true
        schema:
          $ref: '#/definitions/services.CustomerCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: 创建客户
      tags:
      - 客户管理
  /api/customers/{id}:
    get:
      consumes:
      - application/json
      description: 根据ID获取客户的详细信息
      parameters:
      - description: 客户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: 获取客户详情
      tags:
      - 客户管理
    put:
      consumes:
      - application/json
      description: 更新客户的基本信息和扩展信息
      parameters:
      - description: 客户ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新信息
        in: body
        name: customer
        required: true
        schema:
          $ref: '#/definitions/services.CustomerUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: 更新客户信息
      tags:
      - 客户管理
  /api/customers/{id}/activity:
    get:
      consumes:
      - application/json
      description: 获取客户的最近活动记录，包括会话、工单等
      parameters:
      - description: 客户ID
        in: path
        name: id
        required: true
        type: integer
      - description: 记录数量限制
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.CustomerActivity'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: 获取客户活动记录
      tags:
      - 客户管理
  /api/customers/{id}/notes:
    post:
      consumes:
      - application/json
      description: 为客户添加备注信息
      parameters:
      - description: 客户ID
        in: path
        name: id
        required: true
        type: integer
      - description: 备注信息
        in: body
        name: note
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: 添加客户备注
      tags:
      - 客户管理
  /api/customers/{id}/revoke-tokens:
    post:
      consumes:
      - application/json
      description: 提升 token_version 并刷新 token_valid_after，使该客户旧 token 失效
      parameters:
      - description: 客户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: 强制失效客户 token
      tags:
      - 客户管理
  /api/customers/{id}/tags:
    put:
      consumes:
      - application/json
      description: 更新客户的标签信息
      parameters:
      - description: 客户ID
        in: path
        name: id
        required: true
        type: integer
      - description: 标签信息
        in: body
        name: tags
        required: true
        schema:
          additionalProperties:
            items:
              type: string
            type: array
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: 更新客户标签
      tags:
      - 客户管理
  /api/customers/stats:
    get:
      consumes:
      - application/json
      description: 获取客户相关的统计数据
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.CustomerStats'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: 获取客户统计
      tags:
      - 客户管理
...
swagger: "2.0"